refactor (config/config): added new defaults. -> added default_platform_image. -> added return_default_value function. -> added other validations. refactor (config/default): added new defaults. fix (models/assessments): edited duration and retakes data types. feat (rules/assessments): added rules for assessments actions. fix (rules/notifications): -> removed withMessage callback from custom checks where unnecessary. fix (rules/platformNotifications): -> removed withMessage callback from custom checks where unnecessary. fix (rules/platforms): -> removed withMessage callback from custom checks where unnecessary. fix (rules/platformUsers): -> removed withMessage callback from custom checks where unnecessary. fix (rules/users): -> removed withMessage callback from custom checks where unnecessary.

+++ app/models/assessments.model.js
@@ -88,11 +88,11 @@ export default (sequelize, Sequelize) => {
             allowNull: true,
         },
         duration: {
-            type: Sequelize.INTEGER(4),
+            type: Sequelize.INTEGER(5),
             allowNull: true,
         },
         retakes: {
-            type: Sequelize.INTEGER(4),
+            type: Sequelize.INTEGER(3),
             allowNull: true,
         },
         status: {
